+++ inst/include/simmer/process/arrival.h
@@ -67,13 +67,15 @@ namespace simmer {
     Arrival(Simulator* sim, const std::string& name, int mon, Order order,
             Activity* first_activity, int priority = 0, Source* src = NULL)
       : Process(sim, name, mon, priority), order(order), src(src), paused(0),
-        clones(new int(0)), activity(first_activity), timer(NULL),
-        dropout(NULL), batch(NULL), act_shd(new ActVec()) { init(); }
+        sync(new Arrival*(NULL)), clones(new int(0)), activity(first_activity),
+        timer(NULL), dropout(NULL), batch(NULL), act_shd(new ActVec())
+    { init(); }
 
     Arrival(const Arrival& o)
-      : Process(o), order(o.order), src(o.src), paused(o.paused),
-        clones(o.clones), activity(NULL), attributes(o.attributes),
-        timer(NULL), dropout(NULL), batch(NULL), act_shd(o.act_shd) { init(); }
+      : Process(o), order(o.order), src(o.src), paused(o.paused), sync(o.sync),
+        clones(o.clones), activity(NULL), attributes(o.attributes), timer(NULL),
+        dropout(NULL), batch(NULL), act_shd(o.act_shd)
+    { init(); *sync = NULL; }
 
     ~Arrival() { reset(); }
 
@@ -142,7 +144,13 @@ namespace simmer {
 
     virtual size_t size() const { Rcpp::stop("'%s' is not a batch", name); }
 
-    int get_clones() const { return *clones; }
+    bool sync_keep(bool wait) {
+      if (*sync == NULL && (*clones == 1 || !wait))
+        *sync = this;
+      if (*sync == this)
+        return true;
+      return false;
+    }
 
     virtual void set_attribute(const std::string& key, double value, bool global=false) {
       if (global) return sim->set_attribute(key, value);
@@ -235,6 +243,7 @@ namespace simmer {
   private:
     Source* src;
     int paused;
+    Arrival** sync;
     int* clones;          /**< number of active clones */
     ArrStatus status;     /**< arrival timing status */
     ArrTime lifetime;     /**< time spent in the whole trajectory */
@@ -265,6 +274,7 @@ namespace simmer {
         foreach_ (ActVec::value_type& itr, *act_shd)
           itr->remove(this);
         delete act_shd;
+        delete sync;
         delete clones;
       }
       sim->unregister_arrival(this);